+++ core/src/main/java/com/graphhopper/routing/util/RacingBikeFlagEncoder.java
@@ -136,7 +136,7 @@ public class RacingBikeFlagEncoder extends BikeCommonFlagEncoder
     }
 
     @Override
-    public short getVersion()
+    public int getVersion()
     {
         return 1;
     }
